+++ src/components/Page.tsx
@@ -19,7 +19,7 @@ import { createProjectKeys, inviteUserKeys, createGoalKeys, createTeamKeys } fro
 
 import { Theme } from './Theme';
 import { GlobalStyle } from './GlobalStyle';
-import { Header } from './Header';
+import { PageHeader } from './PageHeader';
 
 const ModalOnEvent = dynamic(() => import('./ModalOnEvent'));
 const ProjectCreateForm = dynamic(() => import('./ProjectCreateForm'));
@@ -86,7 +86,7 @@ export const Page: React.FC<PageProps> = ({ user, ssrTime, title = 'Untitled', l
                 position="bottom-right"
             />
 
-            <Header />
+            <PageHeader />
 
             <StyledContent>{children}</StyledContent>
