+++ src/components/TeamListItem.tsx
@@ -3,12 +3,11 @@ import Link from 'next/link';
 
 import { nullable } from '@common/utils/nullable';
 import { Text } from '@common/Text';
+import { UserPic } from '@common/UserPic';
 
 import { routes } from '../hooks/router';
 import type { Team } from '../../graphql/@generated/genql';
 import { gray4, textColor, gray10, gapM, gapS, gray7 } from '../design/@generated/themes';
-
-import { UserPic } from './UserPic';
 
 interface TeamListItemProps {
     team: Team;
