+++ src/components/lib/TaskanyLogo.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { brandColor, textColor } from '../design/@generated/themes';
+import { brandColor, textColor } from '../../design/@generated/themes';
 
 const sizesMap = {
     s: 24,
